Recreate sdk each time
